published/hr/backend/requirements.txt
# ============================================================================
# HR Helper - Production Requirements
# ============================================================================
# Этот файл содержит только реально используемые зависимости проекта
# Транзитивные зависимости (зависимости зависимостей) устанавливаются автоматически
# ============================================================================

# ============================================================================
# Django Core
# ============================================================================
Django==4.2.16
asgiref==3.9.2
sqlparse==0.5.3
pytz==2025.2
tzdata==2025.2

# ============================================================================
# Django Extensions & Middleware
# ============================================================================
djangorestframework==3.16.1  # REST API
django-cors-headers==4.9.0  # CORS поддержка
django-allauth==65.11.2  # OAuth аутентификация (Google)
django-redis==6.0.0  # Redis кэширование
django-timezone-field==7.1  # Поля с таймзонами
django-celery-beat==2.8.1  # Планировщик задач Celery

# ============================================================================
# Database
# ============================================================================
psycopg2-binary==2.9.9  # PostgreSQL драйвер (для продакшена)

# ============================================================================
# Celery & Redis (фоновые задачи)
# ============================================================================
celery==5.5.3
redis==6.4.0
kombu==5.5.4  # Celery message transport
vine==5.1.0  # Celery dependency
billiard==4.2.2  # Celery multiprocessing
amqp==5.3.1  # Celery AMQP support

# ============================================================================
# Google API (OAuth, Calendar, Drive, Sheets)
# ============================================================================
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-core==2.25.1
googleapis-common-protos==1.70.0
httplib2==0.31.0
uritemplate==4.2.0

# ============================================================================
# HTTP & Networking
# ============================================================================
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# ============================================================================
# Authentication & Security
# ============================================================================
PyJWT==2.10.1
cryptography==46.0.1
oauthlib==3.3.1
requests-oauthlib==2.0.0

# ============================================================================
# Excel Export (openpyxl используется в reporting и hiring_plan)
# ============================================================================
openpyxl==3.1.5

# ============================================================================
# Environment & Configuration
# ============================================================================
python-dotenv==1.1.1  # Загрузка .env файлов

# ============================================================================
# Production Server
# ============================================================================
gunicorn==21.2.0  # WSGI сервер для продакшена

# ============================================================================
# Utilities
# ============================================================================
python-dateutil==2.9.0.post0  # Парсинг дат
six==1.17.0  # Python 2/3 совместимость
typing_extensions==4.15.0  # Расширения типизации

# ============================================================================
# Protobuf (зависимость Google API)
# ============================================================================
protobuf==6.32.1
proto-plus==1.26.1

# ============================================================================
# Crypto & Encoding (зависимости других библиотек)
# ============================================================================
cffi==2.0.0
pycparser==2.23
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
pyaes==1.6.1

# ============================================================================
# Celery Beat (планировщик задач)
# ============================================================================
python-crontab==3.3.0  # Используется django-celery-beat для парсинга cron
cron_descriptor==2.0.6  # Описание cron выражений

# ============================================================================
# Click (CLI утилиты, используется Celery и другими)
# ============================================================================
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0

# ============================================================================
# Дополнительные утилиты
# ============================================================================
packaging==25.0
pyparsing==3.2.5
cachetools==5.5.2

# ============================================================================
# Image Processing
# ============================================================================
pillow==11.3.0  # Требуется для ImageField

# ============================================================================
# Wiki: WYSIWYG редактор и миграция Markdown
# ============================================================================
django-summernote==0.8.20.0  # Редактор Summernote для вики
markdown==3.7  # Конвертация Markdown → HTML при миграции вики
